PlayerInterface improved
