+++ executors.py
@@ -12,7 +12,6 @@ from dimmers import ShellyDimmer, ShellyProDimmer2PM, TRANSITION_DEFAULT
 from gpiozero import Button as _Button  # type: ignore
 from instruments import BellSet, DrumSet
 from lights import LightSet
-from mode_interface import AutoModeChangeEntry, ModeConstructor
 import modes
 from mode_defs import (
     BuildBrightness,  EvenOddFade, RotateReversible, RandomFade, RapidFade
@@ -31,7 +30,7 @@ from sequences import (
     rotate_sides,
 )
 from signs_song import SignsSong
-from specialparams import DimmerParams, SpecialParams
+from specialparams import DimmerParams, SpecialParams, AutoModeChangeEntry, ModeConstructor
 
 def setup_devices(brightness_factor: float):
     """"""
