+++ specialparams.py
@@ -3,6 +3,8 @@
 from collections.abc import Callable
 from dataclasses import dataclass
 
+from typing import Any, Type
+
 @dataclass
 class SpecialParams:
     """Base class for special parameters."""
@@ -21,3 +23,14 @@ class DimmerParams(SpecialParams):
     speed_factor: float = 1.0
     transition_on: float | None = 0.5  # !!!! None
     transition_off: float | None = 0.5  # !!!! None
+
+@dataclass
+class ModeConstructor:
+    name: str
+    mode_class: Type
+    kwargs: dict[str, Any]
+
+@dataclass
+class AutoModeChangeEntry:
+    duration_seconds: int
+    mode_index: int
